Update patch for mastodon-bird-ui

+++ cloneThemes.sh
@@ -35,9 +35,13 @@ pullOrClone "Cyberpunk-Neon" "https://github.com/Roboron3042/Cyberpunk-Neon.git"
 cp Cyberpunk-Neon/CSS/mastodon-cyberpunk-neon.css ../app/javascript/styles/_cyberpunk-neon.scss
 
 ## - mastodon-bird-ui
+if [ -d mastodon-bird-ui ]; then
+  cd mastodon-bird-ui
+  git reset --hard HEAD^
+  cd ..
+fi
 pullOrClone "mastodon-bird-ui" "https://github.com/ronilaukkarinen/mastodon-bird-ui.git"
 cd mastodon-bird-ui
-git reset --hard HEAD^
 git checkout nightly && git pull
 git apply ../../_fix-mastodon-bird-ui.patch
 bash scripts/install-to-mastodon.sh --path ../..
